Add an EventDispatcher to each Viewport
