+++ packages/core/src/core/viewport.ts
@@ -6,6 +6,7 @@ import { Box2 } from "../math/box2";
 import { vec2, vec3 } from "gl-matrix";
 import { generateUUID } from "../utilities/uuid_generator";
 import { Logger } from "../utilities/logger";
+import { EventContext, EventDispatcher } from "./event_dispatcher";
 
 export interface ViewportConfig {
   id?: string;
@@ -20,6 +21,7 @@ export class Viewport {
   public readonly element: HTMLElement;
   public readonly camera: Camera;
   public readonly layerManager: LayerManager;
+  public readonly events: EventDispatcher;
   public cameraControls?: CameraControls;
 
   constructor(config: ViewportConfig, layerManager: LayerManager) {
@@ -29,6 +31,23 @@ export class Viewport {
     this.layerManager = layerManager;
     this.cameraControls = config.cameraControls;
     this.updateAspectRatio();
+    this.events = new EventDispatcher(this.element);
+    this.events.addEventListener((event: EventContext) => {
+      if (
+        event.event instanceof PointerEvent ||
+        event.event instanceof WheelEvent
+      ) {
+        const { clientX, clientY } = event.event;
+        const client = vec2.fromValues(clientX, clientY);
+        event.clipPos = this.clientToClip(client, 0);
+        event.worldPos = this.camera.clipToWorld(event.clipPos);
+      }
+      for (const layer of this.layerManager.layers) {
+        layer.onEvent(event);
+        if (event.propagationStopped) return;
+      }
+      this.cameraControls?.onEvent(event);
+    });
 
     for (const layer of config.layers ?? []) {
       this.layerManager.add(layer);
